+++ src/main/java/com/postsquad/scoup/web/user/repository/UserRepository.java
@@ -1,13 +1,10 @@
 package com.postsquad.scoup.web.user.repository;
 
-import com.postsquad.scoup.web.user.User;
-import org.springframework.data.jpa.repository.JpaRepository;
+import com.postsquad.scoup.web.user.domain.User;
 import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
-@Repository
 public interface UserRepository extends CrudRepository<User, Long> {
     Optional<User> findByEmail(String email);
 }
